refactor pagination, add hooks
