--- src/Pages/ForPlaers/index.tsx
@@ -1,12 +1,0 @@
-import Table from "../../components/Table";
-
-const ForPlaersPage = () => {
-  return (
-    <div>
-      <h1>Tournaments available nearby</h1>
-      <Table />
-    </div>
-  );
-};
-
-export default ForPlaersPage;
