+++ src/Pages/Tournaments/index.tsx
@@ -1,0 +1,12 @@
+import Pagination from "../../components/Pagination/Pagination";
+
+const Tournaments = () => {
+  return (
+    <div>
+      <h1>Tournaments available nearby</h1>
+      <Pagination />
+    </div>
+  );
+};
+
+export default Tournaments;
