+++ src/components/MainRouter/index.tsx
@@ -1,17 +1,17 @@
 import { RouteObject, useRoutes } from "react-router-dom";
 import AboutPage from "../../Pages/About";
 import ForOrginizerPage from "../../Pages/ForOrganizer";
-import ForPlaersPage from "../../Pages/ForPlaers";
 import HomePage from "../../Pages/Home";
 import TournamentInfo from "../../Pages/TournamentInfo";
+import Tournaments from "../../Pages/Tournaments";
 
 const MainRouter = () => {
   const basedPath: RouteObject[] = [
     { path: "/", element: <HomePage /> },
-    { path: "/forplaers", element: <ForPlaersPage /> },
     { path: "/about", element: <AboutPage /> },
     { path: "/organizer", element: <ForOrginizerPage /> },
-    { path: "/tournament/:id", element: <TournamentInfo /> },
+    { path: "/tournaments", element: <Tournaments /> },
+    { path: "/tournaments/:id", element: <TournamentInfo /> },
   ];
 
   const resultPath: RouteObject[] = basedPath;
