+++ src/hooks/useGetPlayers/type.ts
@@ -1,0 +1,9 @@
+export interface ITournament {
+  id: string;
+  tournament: string;
+  player_name: string;
+  created: string;
+  updated: string;
+}
+
+export type TGetPlayers = [ITournament[], boolean, (node?: Element | null | undefined) => void];
