+++ src/hooks/useGetTournaments/type.ts
@@ -1,0 +1,11 @@
+import React from "react";
+
+export interface ITournament {
+  id: number;
+  tournamentName: string;
+  game: string;
+  location: string;
+  availableSpots: number;
+}
+
+export type TGetTournaments = [ITournament[], number, React.Dispatch<React.SetStateAction<number>>];
